refactor: local환경에서 docker-mysql 사용 및 데이터 초기화를 테스트 전으로 이동
