+++ src/test/resources/clear.sql
@@ -1,0 +1,12 @@
+DELETE from answer;
+DELETE from agreement;
+DELETE from petition;
+
+DELETE from sign_up_verification_info;
+DELETE from password_verification_info;
+DELETE from user;
+
+DELETE from answer_aud;
+DELETE from petition_aud;
+DELETE from revinfo;
+
